Create new activated-extension test suite
