+++ extensions/ql-vscode/jest.config.js
@@ -8,6 +8,7 @@ module.exports = {
   projects: [
     "<rootDir>/src/view",
     "<rootDir>/test/unit-tests",
+    "<rootDir>/test/vscode-tests/activated-extension",
     "<rootDir>/test/vscode-tests/cli-integration",
     "<rootDir>/test/vscode-tests/no-workspace",
     "<rootDir>/test/vscode-tests/minimal-workspace",
