+++ extensions/ql-vscode/package.json
@@ -1330,6 +1330,7 @@
     "test:unit": "cross-env TZ=UTC LANG=en-US jest --projects test/unit-tests",
     "test:view": "jest --projects src/view",
     "integration": "npm-run-all integration:*",
+    "integration:activated-extension": "jest --projects test/vscode-tests/activated-extension",
     "integration:no-workspace": "jest --projects test/vscode-tests/no-workspace",
     "integration:minimal-workspace": "jest --projects test/vscode-tests/minimal-workspace",
     "cli-integration": "jest --projects test/vscode-tests/cli-integration",
