+++ extensions/ql-vscode/test/vscode-tests/activated-extension/databases/db-panel.test.ts
