+++ extensions/ql-vscode/test/vscode-tests/activated-extension/jest.config.ts
