+++ extensions/ql-vscode/test/vscode-tests/activated-extension/jest.setup.ts
@@ -1,0 +1,1 @@
+import "../jest.activated-extension.setup";
