+++ extensions/ql-vscode/test/vscode-tests/cli-integration/new-query.test.ts
@@ -11,7 +11,7 @@ import { describeWithCodeQL } from "../cli";
 import { QueryServerClient } from "../../../src/query-server/queryserver-client";
 import { extLogger, ProgressReporter } from "../../../src/common";
 import { QueryResultType } from "../../../src/pure/new-messages";
-import { cleanDatabases, dbLoc, storagePath } from "./global.helper";
+import { cleanDatabases, dbLoc, storagePath } from "../global.helper";
 import { importArchiveDatabase } from "../../../src/databaseFetcher";
 
 const baseDir = join(__dirname, "../../../test/data");
